requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
httpx==0.27.2
google-generativeai==0.8.3
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2  # also used by TestClient
